# backend/services/wishlist/gen/api.gen.go
// Package wishlist_server provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package wishlist_server

import (
	"context"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// CreateWishlistItemRequest defines model for CreateWishlistItemRequest.
type CreateWishlistItemRequest struct {
	Data map[string]interface{} `json:"data"`
	Type string                 `json:"type"`
}

// CreateWishlistRequest defines model for CreateWishlistRequest.
type CreateWishlistRequest struct {
	Description *string `json:"description"`
	Title       string  `json:"title"`
}

// UpdateWishlistItemRequest defines model for UpdateWishlistItemRequest.
type UpdateWishlistItemRequest struct {
	Data *map[string]interface{} `json:"data,omitempty"`
	Type *string                 `json:"type,omitempty"`
}

// UpdateWishlistRequest defines model for UpdateWishlistRequest.
type UpdateWishlistRequest struct {
	Description *string `json:"description"`
	Title       *string `json:"title,omitempty"`
}

// Wishlist defines model for Wishlist.
type Wishlist struct {
	CreatedAt   time.Time          `json:"createdAt"`
	Description *string            `json:"description"`
	Id          openapi_types.UUID `json:"id"`
	Items       []WishlistItem     `json:"items"`
	Title       string             `json:"title"`
	UpdatedAt   time.Time          `json:"updatedAt"`
	UserId      openapi_types.UUID `json:"userId"`
}

// WishlistItem defines model for WishlistItem.
type WishlistItem struct {
	CreatedAt *time.Time `json:"createdAt,omitempty"`

	// Data Item-specific payload
	Data map[string]interface{} `json:"data"`
	Id   openapi_types.UUID     `json:"id"`

	// Type Discriminator for item type (e.g., "text", "marketplace")
	Type      string     `json:"type"`
	UpdatedAt *time.Time `json:"updatedAt,omitempty"`
}

// PostWishlistsJSONRequestBody defines body for PostWishlists for application/json ContentType.
type PostWishlistsJSONRequestBody = CreateWishlistRequest

// PutWishlistsWishlistIdJSONRequestBody defines body for PutWishlistsWishlistId for application/json ContentType.
type PutWishlistsWishlistIdJSONRequestBody = UpdateWishlistRequest

// PostWishlistsWishlistIdItemsJSONRequestBody defines body for PostWishlistsWishlistIdItems for application/json ContentType.
type PostWishlistsWishlistIdItemsJSONRequestBody = CreateWishlistItemRequest

// PutWishlistsWishlistIdItemsItemIdJSONRequestBody defines body for PutWishlistsWishlistIdItemsItemId for application/json ContentType.
type PutWishlistsWishlistIdItemsItemIdJSONRequestBody = UpdateWishlistItemRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// List wishlists of the authenticated user
	// (GET /wishlists)
	GetWishlists(w http.ResponseWriter, r *http.Request)
	// Create a new wishlist for the authenticated user
	// (POST /wishlists)
	PostWishlists(w http.ResponseWriter, r *http.Request)
	// Delete a wishlist (owner only)
	// (DELETE /wishlists/{wishlistId})
	DeleteWishlistsWishlistId(w http.ResponseWriter, r *http.Request, wishlistId openapi_types.UUID)
	// Get a wishlist by ID (public endpoint)
	// (GET /wishlists/{wishlistId})
	GetWishlistsWishlistId(w http.ResponseWriter, r *http.Request, wishlistId openapi_types.UUID)
	// Update a wishlist (owner only)
	// (PUT /wishlists/{wishlistId})
	PutWishlistsWishlistId(w http.ResponseWriter, r *http.Request, wishlistId openapi_types.UUID)
	// Add an item to a wishlist (owner only)
	// (POST /wishlists/{wishlistId}/items)
	PostWishlistsWishlistIdItems(w http.ResponseWriter, r *http.Request, wishlistId openapi_types.UUID)
	// Remove an item from a wishlist (owner only)
	// (DELETE /wishlists/{wishlistId}/items/{itemId})
	DeleteWishlistsWishlistIdItemsItemId(w http.ResponseWriter, r *http.Request, wishlistId openapi_types.UUID, itemId openapi_types.UUID)
	// Update a wishlist item (owner only)
	// (PUT /wishlists/{wishlistId}/items/{itemId})
	PutWishlistsWishlistIdItemsItemId(w http.ResponseWriter, r *http.Request, wishlistId openapi_types.UUID, itemId openapi_types.UUID)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// List wishlists of the authenticated user
// (GET /wishlists)
func (_ Unimplemented) GetWishlists(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Create a new wishlist for the authenticated user
// (POST /wishlists)
func (_ Unimplemented) PostWishlists(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Delete a wishlist (owner only)
// (DELETE /wishlists/{wishlistId})
func (_ Unimplemented) DeleteWishlistsWishlistId(w http.ResponseWriter, r *http.Request, wishlistId openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get a wishlist by ID (public endpoint)
// (GET /wishlists/{wishlistId})
func (_ Unimplemented) GetWishlistsWishlistId(w http.ResponseWriter, r *http.Request, wishlistId openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Update a wishlist (owner only)
// (PUT /wishlists/{wishlistId})
func (_ Unimplemented) PutWishlistsWishlistId(w http.ResponseWriter, r *http.Request, wishlistId openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Add an item to a wishlist (owner only)
// (POST /wishlists/{wishlistId}/items)
func (_ Unimplemented) PostWishlistsWishlistIdItems(w http.ResponseWriter, r *http.Request, wishlistId openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Remove an item from a wishlist (owner only)
// (DELETE /wishlists/{wishlistId}/items/{itemId})
func (_ Unimplemented) DeleteWishlistsWishlistIdItemsItemId(w http.ResponseWriter, r *http.Request, wishlistId openapi_types.UUID, itemId openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Update a wishlist item (owner only)
// (PUT /wishlists/{wishlistId}/items/{itemId})
func (_ Unimplemented) PutWishlistsWishlistIdItemsItemId(w http.ResponseWriter, r *http.Request, wishlistId openapi_types.UUID, itemId openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetWishlists operation middleware
func (siw *ServerInterfaceWrapper) GetWishlists(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetWishlists(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostWishlists operation middleware
func (siw *ServerInterfaceWrapper) PostWishlists(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostWishlists(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeleteWishlistsWishlistId operation middleware
func (siw *ServerInterfaceWrapper) DeleteWishlistsWishlistId(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "wishlistId" -------------
	var wishlistId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "wishlistId", chi.URLParam(r, "wishlistId"), &wishlistId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "wishlistId", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteWishlistsWishlistId(w, r, wishlistId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetWishlistsWishlistId operation middleware
func (siw *ServerInterfaceWrapper) GetWishlistsWishlistId(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "wishlistId" -------------
	var wishlistId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "wishlistId", chi.URLParam(r, "wishlistId"), &wishlistId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "wishlistId", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetWishlistsWishlistId(w, r, wishlistId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PutWishlistsWishlistId operation middleware
func (siw *ServerInterfaceWrapper) PutWishlistsWishlistId(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "wishlistId" -------------
	var wishlistId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "wishlistId", chi.URLParam(r, "wishlistId"), &wishlistId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "wishlistId", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PutWishlistsWishlistId(w, r, wishlistId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PostWishlistsWishlistIdItems operation middleware
func (siw *ServerInterfaceWrapper) PostWishlistsWishlistIdItems(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "wishlistId" -------------
	var wishlistId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "wishlistId", chi.URLParam(r, "wishlistId"), &wishlistId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "wishlistId", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PostWishlistsWishlistIdItems(w, r, wishlistId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeleteWishlistsWishlistIdItemsItemId operation middleware
func (siw *ServerInterfaceWrapper) DeleteWishlistsWishlistIdItemsItemId(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "wishlistId" -------------
	var wishlistId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "wishlistId", chi.URLParam(r, "wishlistId"), &wishlistId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "wishlistId", Err: err})
		return
	}

	// ------------- Path parameter "itemId" -------------
	var itemId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "itemId", chi.URLParam(r, "itemId"), &itemId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "itemId", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteWishlistsWishlistIdItemsItemId(w, r, wishlistId, itemId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PutWishlistsWishlistIdItemsItemId operation middleware
func (siw *ServerInterfaceWrapper) PutWishlistsWishlistIdItemsItemId(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "wishlistId" -------------
	var wishlistId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "wishlistId", chi.URLParam(r, "wishlistId"), &wishlistId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "wishlistId", Err: err})
		return
	}

	// ------------- Path parameter "itemId" -------------
	var itemId openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "itemId", chi.URLParam(r, "itemId"), &itemId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "itemId", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PutWishlistsWishlistIdItemsItemId(w, r, wishlistId, itemId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/wishlists", wrapper.GetWishlists)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/wishlists", wrapper.PostWishlists)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/wishlists/{wishlistId}", wrapper.DeleteWishlistsWishlistId)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/wishlists/{wishlistId}", wrapper.GetWishlistsWishlistId)
	})
	r.Group(func(r chi.Router) {
		r.Put(options.BaseURL+"/wishlists/{wishlistId}", wrapper.PutWishlistsWishlistId)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/wishlists/{wishlistId}/items", wrapper.PostWishlistsWishlistIdItems)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/wishlists/{wishlistId}/items/{itemId}", wrapper.DeleteWishlistsWishlistIdItemsItemId)
	})
	r.Group(func(r chi.Router) {
		r.Put(options.BaseURL+"/wishlists/{wishlistId}/items/{itemId}", wrapper.PutWishlistsWishlistIdItemsItemId)
	})

	return r
}